pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "atlas-stock-ml"
version = "1.0.0"
description = "AI-powered stock trading dashboard with ML price predictions"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [{name = "Anh Dang"}]

dependencies = [
    "yfinance>=0.2.40",
    "streamlit>=1.38",
    "pandas>=2.1",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "plotly>=5.20",
    "matplotlib>=3.8",
    "requests>=2.31",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "pandera>=0.18",
]

[project.optional-dependencies]
gcp = [
    "google-cloud-aiplatform>=1.60",
    "google-cloud-bigquery>=3.20",
    "google-cloud-storage>=2.14",
    "google-cloud-secret-manager>=2.18",
    "google-auth>=2.25",
    "db-dtypes>=1.2",
]
ml = [
    "tensorflow>=2.13,<2.16",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-mock>=3.12",
    "flake8>=7.0",
    "black>=24.0",
    "mypy>=1.8",
    "pre-commit>=3.6",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "network: marks tests that require network access",
    "gcp: marks tests that require GCP credentials",
    "tensorflow: marks tests that require TensorFlow",
]

[tool.coverage.run]
source = ["ml", "data", "ui"]
omit = ["tests/*", "gcp/*", "*/migrations/*"]

[tool.coverage.report]
fail_under = 60
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.",
    "if HAS_TENSORFLOW",
    "if TYPE_CHECKING",
    "raise NotImplementedError",
]

[tool.black]
line-length = 120
target-version = ["py39", "py310", "py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
check_untyped_defs = true

[tool.flake8]
max-line-length = 120
extend-ignore = ["E203", "W503"]
exclude = [".git", "__pycache__", "venv", "archive"]

[tool.setuptools.packages.find]
include = ["ml*", "data*", "ui*", "config*"]
